Refactor db get events and create event code to follow new project structure
