+++ functions/lambda/templates/pages/home.templ
@@ -5,7 +5,7 @@ import (
 	"github.com/meetnearme/api/functions/lambda/helpers"
 )
 
-templ HomePage(events []services.Event) {
+templ HomePage(events []services.EventSelect) {
 	<div class="header-hero">
 		<h1 class="text-2xl md:text-5xl text-center mt-5">
 			Magically <span class="font-system">&#x2728</span> collect in-person events you care about <span class="font-system">&#x1F938</span> and discover new ones
