+++ functions/lambda/templates/pages/home_templ.go
@@ -15,7 +15,7 @@ import (
 	"github.com/meetnearme/api/functions/lambda/services"
 )
 
-func HomePage(events []services.Event) templ.Component {
+func HomePage(events []services.EventSelect) templ.Component {
 	return templ.ComponentFunc(func(ctx context.Context, templ_7745c5c3_W io.Writer) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_Buffer, templ_7745c5c3_IsBuffer := templ_7745c5c3_W.(*bytes.Buffer)
 		if !templ_7745c5c3_IsBuffer {
